Convert web/src/api/locations-api.js to TypeScript
- Convert JSDoc type definitions to TypeScript types
- Use WhereOptions, FiltersOptions, QueryOptions helper types from base-api
- Export all type definitions for reuse
- Add proper type annotations to all methods
- Preserve debounceWithArgsCache usage with proper typing
